test(notebook): characterize shell process behavior
